+++ client/src/components/shared/inputs/TextInputComponent.vue
@@ -22,7 +22,8 @@ export default {
     const store = useTestStore();
 
     function updateValue(value) {
-      store.updateField({ field: props.field.name, value });
+      console.log({ field: props.field.name, value });
+      store.handleFieldUpdate({ field: props.field.name, value });
     }
 
     return { updateValue };
